refactor(test): improve random test import and split importer class
